refactor frontend chat component
